+++ crates/zizmor/src/audit/concurrency_limits.rs
@@ -6,7 +6,7 @@ use crate::{
     state::AuditState,
 };
 use anyhow::Result;
-use github_actions_models::{common::expr::BoE, workflow::Concurrency};
+use github_actions_models::workflow::Concurrency;
 
 pub(crate) struct ConcurrencyLimits;
 
@@ -28,29 +28,6 @@ impl Audit for ConcurrencyLimits {
     ) -> Result<Vec<Finding<'doc>>> {
         let mut findings = vec![];
         match &workflow.concurrency {
-            Some(Concurrency::Rich {
-                group: _,
-                cancel_in_progress,
-            }) => {
-                if let BoE::Literal(cancel) = &cancel_in_progress
-                    && !cancel
-                {
-                    findings.push(
-                        Self::finding()
-                            .confidence(Confidence::High)
-                            .severity(Severity::Low)
-                            .persona(Persona::Pedantic)
-                            .add_location(
-                                workflow
-                                    .location()
-                                    .primary()
-                                    .with_keys(["concurrency".into()])
-                                    .annotated("cancel-in-progress set to false"),
-                            )
-                            .build(workflow)?,
-                    );
-                };
-            }
             Some(Concurrency::Bare(_)) => {
                 findings.push(
                     Self::finding()
@@ -82,6 +59,11 @@ impl Audit for ConcurrencyLimits {
                         .build(workflow)?,
                 );
             }
+            // NOTE: Per #1302, we don't nag the user if they've explicitly set
+            // `cancel-in-progress: false` or similar. This is like with the
+            // artipacked audit, where `persist-credentials: true` is seen as
+            // a positive signal of user intent.
+            _ => {}
         }
 
         Ok(findings)